pyproject.toml
[project]
name = "cafe"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "django-rest-framework>=0.1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "pillow>=11.2.1",
    "ruff>=0.11.12",
]
